Removed remote input code that is no longer necessary after new SampleBuffer updates.  Also continued to move common code into the NetInput class, and made some minor code clean-up changes.
git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler@7823 08b38cba-cd3b-11de-854e-f91c5b6e4272

+++ modules/gadgeteer/gadget/RemoteInputManager/NetUtils.h
@@ -87,7 +87,7 @@ inline float vj_htonf(float f)
 inline int binaryToInt(const char* a, const int len){
     int return_int = 0;
     for (int i = 0; i < len; i++) {
-        return_int += a[i] * pow((double)256,i);  // that means += a[i] * 256 ^ i
+        return_int += (int)(a[i]) * (int)(pow((double)256,i));  // that means += a[i] * 256 ^ i
     }
     return return_int;
 }
